refactor: clean up config
