+++ .env.example
@@ -9,3 +9,5 @@ EMAIL_USER=user@example.com
 EMAIL_PASS=your_email_password
 EMAIL_FROM=user@example.com
 EMAIL_TO=recipient@example.com
+
+ALPHA_VANTAGE_API_KEY=asdf1234qwer5678
